services/inventory/requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
opentelemetry-api==1.31.1
opentelemetry-sdk==1.31.1
opentelemetry-exporter-otlp-proto-http==1.31.1
opentelemetry-instrumentation-fastapi==0.52b1
